class File #8
## 배운 내용
- class 파일 분리
