+++ CMakeLists.txt
@@ -3,4 +3,4 @@ project(cpp_study)
 
 set(CMAKE_CXX_STANDARD 14)
 
-add_executable(cpp_study main.cpp common.cpp common.h)
+add_executable(cpp_study main.cpp common.cpp common.h Character.cpp Character.h)
