+++ cmake-build-debug/CMakeFiles/cpp_study.dir/CXX.includecache
@@ -6,6 +6,10 @@
 
 #IncludeRegexTransform: 
 
+/Users/kimjooyoun/CLionProjects/cpp_study/Character.h
+iostream
+-
+
 /Users/kimjooyoun/CLionProjects/cpp_study/common.h
 
 /Users/kimjooyoun/CLionProjects/cpp_study/main.cpp
@@ -13,4 +17,6 @@ iostream
 -
 common.h
 /Users/kimjooyoun/CLionProjects/cpp_study/common.h
+Character.h
+/Users/kimjooyoun/CLionProjects/cpp_study/Character.h
