+++ cmake-build-debug/CMakeFiles/cpp_study.dir/build.make
@@ -92,19 +92,34 @@ CMakeFiles/cpp_study.dir/common.cpp.s: cmake_force
 	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --green "Compiling CXX source to assembly CMakeFiles/cpp_study.dir/common.cpp.s"
 	/Library/Developer/CommandLineTools/usr/bin/c++ $(CXX_DEFINES) $(CXX_INCLUDES) $(CXX_FLAGS) -S /Users/kimjooyoun/CLionProjects/cpp_study/common.cpp -o CMakeFiles/cpp_study.dir/common.cpp.s
 
+CMakeFiles/cpp_study.dir/Character.cpp.o: CMakeFiles/cpp_study.dir/flags.make
+CMakeFiles/cpp_study.dir/Character.cpp.o: ../Character.cpp
+	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --green --progress-dir=/Users/kimjooyoun/CLionProjects/cpp_study/cmake-build-debug/CMakeFiles --progress-num=$(CMAKE_PROGRESS_3) "Building CXX object CMakeFiles/cpp_study.dir/Character.cpp.o"
+	/Library/Developer/CommandLineTools/usr/bin/c++ $(CXX_DEFINES) $(CXX_INCLUDES) $(CXX_FLAGS) -o CMakeFiles/cpp_study.dir/Character.cpp.o -c /Users/kimjooyoun/CLionProjects/cpp_study/Character.cpp
+
+CMakeFiles/cpp_study.dir/Character.cpp.i: cmake_force
+	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --green "Preprocessing CXX source to CMakeFiles/cpp_study.dir/Character.cpp.i"
+	/Library/Developer/CommandLineTools/usr/bin/c++ $(CXX_DEFINES) $(CXX_INCLUDES) $(CXX_FLAGS) -E /Users/kimjooyoun/CLionProjects/cpp_study/Character.cpp > CMakeFiles/cpp_study.dir/Character.cpp.i
+
+CMakeFiles/cpp_study.dir/Character.cpp.s: cmake_force
+	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --green "Compiling CXX source to assembly CMakeFiles/cpp_study.dir/Character.cpp.s"
+	/Library/Developer/CommandLineTools/usr/bin/c++ $(CXX_DEFINES) $(CXX_INCLUDES) $(CXX_FLAGS) -S /Users/kimjooyoun/CLionProjects/cpp_study/Character.cpp -o CMakeFiles/cpp_study.dir/Character.cpp.s
+
 # Object files for target cpp_study
 cpp_study_OBJECTS = \
 "CMakeFiles/cpp_study.dir/main.cpp.o" \
-"CMakeFiles/cpp_study.dir/common.cpp.o"
+"CMakeFiles/cpp_study.dir/common.cpp.o" \
+"CMakeFiles/cpp_study.dir/Character.cpp.o"
 
 # External object files for target cpp_study
 cpp_study_EXTERNAL_OBJECTS =
 
 cpp_study: CMakeFiles/cpp_study.dir/main.cpp.o
 cpp_study: CMakeFiles/cpp_study.dir/common.cpp.o
+cpp_study: CMakeFiles/cpp_study.dir/Character.cpp.o
 cpp_study: CMakeFiles/cpp_study.dir/build.make
 cpp_study: CMakeFiles/cpp_study.dir/link.txt
-	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --green --bold --progress-dir=/Users/kimjooyoun/CLionProjects/cpp_study/cmake-build-debug/CMakeFiles --progress-num=$(CMAKE_PROGRESS_3) "Linking CXX executable cpp_study"
+	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --green --bold --progress-dir=/Users/kimjooyoun/CLionProjects/cpp_study/cmake-build-debug/CMakeFiles --progress-num=$(CMAKE_PROGRESS_4) "Linking CXX executable cpp_study"
 	$(CMAKE_COMMAND) -E cmake_link_script CMakeFiles/cpp_study.dir/link.txt --verbose=$(VERBOSE)
 
 # Rule to build all files generated by this target.
