+++ cmake-build-debug/CMakeFiles/cpp_study.dir/cmake_clean.cmake
@@ -1,4 +1,5 @@
 file(REMOVE_RECURSE
+  "CMakeFiles/cpp_study.dir/Character.cpp.o"
   "CMakeFiles/cpp_study.dir/common.cpp.o"
   "CMakeFiles/cpp_study.dir/main.cpp.o"
   "cpp_study"
