+++ cmake-build-debug/CMakeFiles/cpp_study.dir/depend.internal
@@ -1,9 +1,13 @@
 # CMAKE generated file: DO NOT EDIT!
 # Generated by "Unix Makefiles" Generator, CMake Version 3.20
 
+CMakeFiles/cpp_study.dir/Character.cpp.o
+ /Users/kimjooyoun/CLionProjects/cpp_study/Character.cpp
+ /Users/kimjooyoun/CLionProjects/cpp_study/Character.h
 CMakeFiles/cpp_study.dir/common.cpp.o
  /Users/kimjooyoun/CLionProjects/cpp_study/common.cpp
  /Users/kimjooyoun/CLionProjects/cpp_study/common.h
 CMakeFiles/cpp_study.dir/main.cpp.o
+ /Users/kimjooyoun/CLionProjects/cpp_study/Character.h
  /Users/kimjooyoun/CLionProjects/cpp_study/common.h
  /Users/kimjooyoun/CLionProjects/cpp_study/main.cpp
