+++ cmake-build-debug/CMakeFiles/cpp_study.dir/depend.make
@@ -1,9 +1,13 @@
 # CMAKE generated file: DO NOT EDIT!
 # Generated by "Unix Makefiles" Generator, CMake Version 3.20
 
+CMakeFiles/cpp_study.dir/Character.cpp.o: \
+ ../Character.cpp \
+ ../Character.h
 CMakeFiles/cpp_study.dir/common.cpp.o: \
  ../common.cpp \
  ../common.h
 CMakeFiles/cpp_study.dir/main.cpp.o: \
+ ../Character.h \
  ../common.h \
  ../main.cpp
