+++ cmake-build-debug/Testing/Temporary/LastTest.log
@@ -1,3 +1,3 @@
-Start testing: Nov 14 15:52 KST
+Start testing: Nov 14 15:57 KST
 ----------------------------------------------------------
-End testing: Nov 14 15:52 KST
+End testing: Nov 14 15:57 KST
