+++ main.cpp
@@ -1,95 +1,9 @@
-// 심화 객체 접근 한정자 -> Character -> this pointer
+// 심화 : class 파일로 분리
 #include <iostream>
 #include "common.h"
-
+#include "Character.h"
 using namespace std;
 
-class Character {
-public:
-    Character();
-    explicit Character(int hp);
-
-public:
-    int getHp() const;
-
-    void setHp(int hp);
-
-    int getSpeed() const;
-
-    void setSpeed(int speed);
-
-    const string &getName() const;
-
-    void setName(const string &name);
-
-private:
-    int hp{};
-    int speed{};
-    string name;
-
-public:
-
-    void move(int dir);
-
-    void jump();
-
-    void hit();
-
-    void speedUp();
-
-    void obtainItem();
-};
-Character::Character(int _hp){
-    hp = _hp;
-}
-Character::Character(){
-    hp = 100;
-}
-
-
-void Character::speedUp() {
-    speed++;
-}
-
-void Character::obtainItem() {
-    speedUp();
-}
-
-int Character::getHp() const {
-    return hp;
-}
-
-void Character::setHp(int hp) {
-    this -> hp = hp;
-}
-
-int Character::getSpeed() const {
-    return speed;
-}
-
-void Character::setSpeed(int speed) {
-    this -> speed = speed;
-}
-
-const string &Character::getName() const {
-    return name;
-}
-
-void Character::setName(const string &name) {
-    this -> name = name;
-}
-
-void Character::jump() {
-
-}
-
-void Character::move(int dir) {
-
-}
-
-void Character::hit() {
-
-}
 
 int main() {
     Character mario;
